Getting closer to having a ClipList
Refactored GrafModel so each instance can return its state as a list of
JSON ops. These ops will be the payload in the "add clip" operation on
a ClipList, so the cliplist can be stored, loaded, and synced
cross-tab. What I go through to implement cut-n-paste for structured
objects, wow.