Level creation and some planning
Now each level.js file only sets a key/value pair in window.vorpLevels,
and the game HTML uses it to generate a playable Vorp. This is prep for
the level editor system to be able to pull in all existing levels, for
testing.
Also did a bit of level editor planning.