urlSquisher2 core code is good enough
Using a static dictionary plus Lempel-Ziv, and paying the cost of a URL
expander prefix and conversion to base64 (quite a hit!) I still get
about 50% compression for my sample URLs.
My own "best substitution" method sometimes outperforms Lempel-Ziv, but
not on average, and mine is much slower to encode, so I've turned it
off.