Zombie 0.3 - Lethal
Touch from a zombie will turn player into another zombie over a
terrifying two second period, during which the player changes color and
loses motor control. When the zombification is complete, a new player
body is assembled.
Refactored the exploding kind of player death to allow for this new
peril.
Zombies are unkillable, but not for long...